Seperate Platform connecting part from BotItem.js
